+++ CHANGES.md
@@ -4,6 +4,13 @@
 ### Changed
 - Support escaping on service param names
 
+## [2.0.0] - 2015-05-27
+
+- Changed namespace from Gorkaio to Gorka
+- CHANGELONG.md renamed to CHANGES.md
+- Removed phpddocumentor and phpcs dev dependencies
+- Moved phpunit.xml to project root
+
 ## [1.0.2] - 2015-02-01
 ### Changed
 - Fixed travis config